backend/pyproject.toml
[project]
name = "chess-polling-game"
version = "0.1.0"
requires-python = ">=3.8"
authors = [{ name="Sergio Pires", email="sergiodanpires@gmail.com" }]
description = "A chess polling game backend."
keywords = ["python", "chess", "voting"]
classifiers = [
  "Programming Language :: Python"
]

dependencies = [
  "flask==3.0.3",
  "flask-restx==1.3.0",
  "Flask-Cors==4.0.0",
  "gunicorn==22.0.0",
  "requests==2.31.0",
  "gevent==24.2.1",
  "redis==5.0.4",
  "chess==1.10.0",
  "python-dotenv==1.0.0",
  "pymongo==4.7.1",
  "APScheduler==3.6.3",
  "tzlocal==2.1"
]
